+++ pc-erp/src/main/java/net/caiban/pc/erp/service/sys/SysAppService.java
@@ -13,7 +13,7 @@ import net.caiban.pc.erp.domain.sys.SysApp;
  */
 public interface SysAppService {
 
-	public SysApp queryByDomain(Integer cid, String domain);
+	public SysApp queryByDomain(Long cid, String domain);
 	
 	public SysApp saveOrUpdate(SysApp app);
